+++ package.json
@@ -23,6 +23,7 @@
     "modern-normalize": "^2.0.0",
     "normalize.css": "^8.0.1",
     "notiflix": "^3.2.7",
+    "rc-scrollbars": "^1.1.6",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-i18next": "^14.1.0",
