+++ src/components/Home/CustomScrollbars/CustomScrollbars.jsx
@@ -1,0 +1,14 @@
+import { Scrollbars } from 'rc-scrollbars';
+
+import css from "./CustomScrollbars.module.css"
+
+export const CustomScrollbars = ({ children }) => {
+  return (
+      <Scrollbars autoHeight autoHeightMax={180} classes={{
+        trackVertical: css.trackVertical,
+        thumbVertical: css.thumbVertical,
+      }}>
+          {children}
+      </Scrollbars>
+  )
+}
