+++ src/components/Home/CustomScrollbars/CustomScrollbars.module.css
@@ -1,0 +1,8 @@
+.trackVertical{
+    background-color: var(--secondary-fifth);
+    width: 4px!important;
+}
+
+.thumbVertical {
+    background-color: var(--secondary-third)!important;
+}
