+++ src/components/Home/DailyNorma/DailyNorma.jsx
@@ -2,12 +2,12 @@ import { ModalSubtitle } from '../CommonStyles.styled';
 import { NormaBlock, NormaBlockFooter } from './DailyNorma.styled';
 import { DailyNormaModal } from '../DailyNormaModal/DailyNormaModal';
 import { useState } from 'react';
-// import { useSelector } from "react-redux";
-// import { selectAuthUser } from "../../../redux/auth/authSelectors";
+import { useSelector } from "react-redux";
+import { selectAuthUser } from "../../../redux/auth/authSelectors";
 
 export const DailyNorma = () => {
-  // const { dailyNorma } = useSelector(selectAuthUser);
-  const dailyNorma = 2;
+  const { waterRate } = useSelector(selectAuthUser);
+  const dailyNorma = waterRate / 1000;
   const [isModalOpen, setIsModalOpen] = useState(false);
   const handlerToggleModal = () => {
     setIsModalOpen(!isModalOpen);
@@ -16,7 +16,7 @@ export const DailyNorma = () => {
     <NormaBlock>
       <ModalSubtitle>My daily norma</ModalSubtitle>
       <NormaBlockFooter>
-        <span>{dailyNorma} L</span>
+        <span>{dailyNorma || 0} L</span>
         <button onClick={handlerToggleModal} type="button">
           Edit
         </button>
