+++ src/index.css
@@ -49,7 +49,7 @@ button {
   cursor: pointer;
 }
 
-::-webkit-scrollbar {
+/* ::-webkit-scrollbar {
   width: 4px;
 }
 
@@ -60,9 +60,9 @@ button {
 
 ::-webkit-scrollbar-track {
   background-color: rgba(239, 237, 232, 0.1);
-}
+} */
 
 /* Firefox */
-* {
+/* * {
   scrollbar-color: #9ebbff rgba(239, 237, 232, 0.1);
-}
+} */
